TSIS/TSIS4/requirements.txt
pygame-ce
psycopg[binary]
